requirements.txt
# Interview Prep AI - Dependencies

# UI Framework
flet==0.21.2

# Database
mysql-connector-python==8.3.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pillow==10.2.0

# AI/LLM Providers
openai==1.12.0
anthropic==0.18.1
boto3==1.34.51

# HTTP Clients
requests==2.31.0
httpx==0.27.0

# Security & Encryption
cryptography==42.0.2
python-dotenv==1.0.1

# Utilities
validators==0.22.0
python-dateutil==2.8.2
